+++ CHANGELOG.md
@@ -1,3 +1,22 @@
-## 1.0.0
+# Changelog
 
-- Initial version.
+All notable changes to this project will be documented in this file.
+
+## [1.0.0] - 2025-08-12
+### Added
+- Initial release of `universal_dropdown` Flutter package.
+- Supports single and multiple selection modes.
+- Searchable dropdown list feature.
+- Customizable item widgets and chips.
+- API-driven dropdown example with Dio integration.
+- Basic usage and advanced usage examples included.
+- Screenshot and GIF demo placeholders in README.
+
+### Fixed
+- N/A
+
+### Changed
+- N/A
+
+
+